chore(tui): clippy cleanup, remove dead code warnings
- Core crate: zero warnings (was 22)
- Terminal crate: only dead_code warnings for future stubs remain
- Fixed clamp pattern, prefix stripping, explicit counter loops
- Removed unnecessary parentheses, unused imports
- Fixed test import for MockTransport
- 51 tests passing (39 core + 12 terminal), 7,142 LOC

+++ crates/tui/core/src/animation.rs
@@ -160,7 +160,7 @@ impl AnimationState {
 
         // Error pulse: red tinted ring
         if matches!(self.mode, SubstrateMode::ErrorPulse) {
-            let pulse = ((t * 4.0).sin() * 0.5 + 0.5) as f32;
+            let pulse = (t * 4.0).sin() * 0.5 + 0.5;
             prims.push(ScenePrimitive::Ring {
                 center,
                 radius: 0.15 + pulse * 0.1,
